requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Limiter==3.5.0
PyMySQL==1.1.0
Werkzeug==3.0.1
python-dotenv==1.0.1
cryptography>=42.0.0
